fix up column types, options, and association between VendorIdentifier and Vendor

+++ src/models/dataSupplier.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
     'DataSupplier',
     {
       id: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INT,
         allowNull: false,
         primaryKey: true
       },
